refactor: AgreementsTest 패키지 위치 수정 및 가다듬기
