--- src/test/java/com/gistpetition/api/agreement/AgreementsSupport.java
@@ -1,9 +1,0 @@
-package com.gistpetition.api.agreement;
-
-import com.gistpetition.api.petition.domain.Agreements;
-
-public class AgreementsSupport extends Agreements {
-    public AgreementsSupport() {
-        super();
-    }
-}
